% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Barplot_Accumulative.R
\name{BetterVis_Barplot_Accumulative}
\alias{BetterVis_Barplot_Accumulative}
\title{BetterVis Accumulative Barplot Function}
\usage{
BetterVis_Barplot_Accumulative(
  data,
  x_var,
  y1_var,
  y2_var,
  fill_var,
  display = 3,
  barplot_color = NULL,
  axis_titles_x = "",
  x_label_angle = 0,
  x_label_bold = FALSE,
  axis_title_x_size = 12,
  axis_titles_y1 = "",
  axis_titles_y2 = "",
  y1_label_bold = FALSE,
  y2_label_bold = FALSE,
  axis_title_y1_size = 12,
  axis_title_y2_size = 12,
  title1 = "",
  title2 = "",
  title1_size = NULL,
  title1_bold = NULL,
  title2_size = NULL,
  title2_bold = NULL,
  title_size = 14,
  title_bold = FALSE,
  legend_show = TRUE,
  legend_title = "",
  legend_size = 10,
  legend_position = "bottom",
  legend_row = NULL,
  legend_col = NULL,
  background_color = "white",
  ylim1 = NULL,
  ylim2 = NULL,
  coord_flip = FALSE,
  grid_ncol = 2,
  grid_layout = rbind(c(1, 1, 1, 1, 1, 2, 2), c(1, 1, 1, 1, 1, 2, 2)),
  p1_theme = NULL,
  p2_theme = NULL,
  p1_guide = NULL,
  p2_guide = NULL
)
}
\arguments{
\item{data}{A data frame containing the variables for plotting.}

\item{x_var}{A string specifying the variable for the x-axis, which should be a factor variable.}

\item{y1_var}{A string specifying the numeric variable for the y-axis of the first plot.}

\item{y2_var}{A string specifying the numeric variable for the y-axis of the second plot.}

\item{fill_var}{A string specifying the variable used for fill colors, which should be a factor variable.}

\item{display}{An integer specifying which plot(s) to display: 1 for the first plot, 2 for the second plot, 3 for both plots.}

\item{barplot_color}{A vector specifying the fill colors corresponding to each level of the fill_var factor.}

\item{axis_titles_x}{A string for the x-axis title (default: \code{""}).}

\item{x_label_angle}{Numeric, defines the angle for x-axis labels.}

\item{x_label_bold}{Logical, determines if x-axis labels should be bold (default: \code{FALSE}).}

\item{axis_title_x_size}{Numeric, specifies the font size for the x-axis title.}

\item{axis_titles_y1}{A string for the y-axis title of the first plot.}

\item{axis_titles_y2}{A string for the y-axis title of the second plot.}

\item{y1_label_bold}{Logical, determines if the y-axis labels on the first plot should be bold (default: \code{FALSE}).}

\item{y2_label_bold}{Logical, determines if the y-axis labels on the second plot should be bold (default: \code{FALSE}).}

\item{axis_title_y1_size}{Numeric, specifies the font size for the y-axis title on the first plot.}

\item{axis_title_y2_size}{Numeric, specifies the font size for the y-axis title on the second plot.}

\item{title1}{A string for the title of the first plot.}

\item{title2}{A string for the title of the second plot.}

\item{title1_size}{Numeric, specifies the font size for the first plot title.}

\item{title1_bold}{Logical, determines if the first plot title should be bold.}

\item{title2_size}{Numeric, specifies the font size for the second plot title.}

\item{title2_bold}{Logical, determines if the second plot title should be bold.}

\item{title_size}{Numeric, specifies the font size for plot titles.}

\item{title_bold}{Logical, determines if plot titles should be bold (default: \code{FALSE}).}

\item{legend_show}{Logical, determines if the legend should be displayed (default: \code{TRUE}).}

\item{legend_title}{A string for the legend title.}

\item{legend_size}{Numeric, specifies the font size for legend text.}

\item{legend_position}{A string specifying the position of the legend, default is \code{"bottom"}.}

\item{background_color}{A string for the background color, default is \code{"white"}.}

\item{ylim1}{A numeric vector, adjusts the y-axis limits for the first plot (optional).}

\item{ylim2}{A numeric vector, adjusts the y-axis limits for the second plot (optional).}

\item{coord_flip}{Logical, determines if the coordinates should be flipped (default: \code{FALSE}).}

\item{grid_ncol}{Numeric, specifies the number of columns for arranging plots (default: \code{2}).}

\item{grid_layout}{Numeric matrix, defines the layout for arranging plots.}
}
\value{
A plot or arranged plots according to specified options.
}
\description{
Generates accumulative bar plots to showcase numerical distribution and sample ratios with comprehensive customization and layout options.
}
\details{
Creates accumulative bar plots with extensive customization options, allowing users to visualize both numeric and percentage data across different categories.
The function provides various options for customizing themes, axes, titles, and legend settings.
}
\examples{

library(ggplot2)
library(dplyr)
library(grid)
library(gridExtra)
library(plyr)


##  示例数据1
data("BetterVis_Barplot_Accumulative_example", package = "BetterVis")
df <- BetterVis_Barplot_Accumulative_example
df$celltype <- factor(df$celltype,levels=rev(unique(df$celltype)))
df$Patient <- factor(df$Patient,levels=c("P21","P48","P53","P54","P08")) #指定柱内顺序
df = ddply(df,'celltype',transform,percent=cell_num/sum(cell_num))

barplot_color = c("#ABDDDE", "#FAD510", "#C6CDF7", "#F4B5BD",  "#FAEED1",
"#0A9F9D",  "#005295", "#E6A0C4", "#C52E19", "orange")


## 示例数据2
iris$Group <- rep(paste0("Group", 1:10), times = 15)
iris$Group <-as.factor(iris$Group)
iris$Group2 <- rep(rep(paste0("Color", 1:5), each = 10),3)
iris$Group2 <- sample(iris$Group2)
iris$Group2 <-as.factor(iris$Group2)
iris<-iris\%>\%dplyr::select(Petal.Width,Group2,Group)
iris<-distinct(iris,Group,Group2,.keep_all = TRUE)
iris = ddply(iris,'Group',transform,percent=Petal.Width/sum(Petal.Width))

BetterVis_Barplot_Accumulative(
  data = BetterVis_Barplot_Accumulative_example, x_var = "celltype", y1_var = "cell_num", y2_var = "percent", fill_var = "Patient",
  display = 3, barplot_color = barplot_color,
  ylim1 = c(0, 15000), ylim2 = c(0, 1),
  coord_flip = TRUE
)
BetterVis_Barplot_Accumulative(
  data = df, x_var = "celltype", y1_var = "cell_num", y2_var = "percent", fill_var = "Patient",
  display = 3, barplot_color = barplot_color,
  ylim1 = c(0, 15000), ylim2 = c(0, 1),
  axis_titles_x = "Cell Types", x_label_angle = 0, x_label_bold = TRUE, axis_title_x_size = 14,
  axis_titles_y1 = "Cell Count", axis_titles_y2 = "Sample Ratio",
  y1_label_bold = TRUE,y2_label_bold =  TRUE, axis_title_y1_size = 14,axis_title_y2_size = 14,
  title1 = "Cell Number Distribution", title2 = "Sample Ratio Distribution",
  title_size = 10,title_bold = TRUE,
  legend_show = TRUE, legend_title = "Legend", legend_size = 14, legend_position = "bottom", legend_row = NULL,legend_col=NULL,
  background_color = "white",
  coord_flip = TRUE,
  grid_ncol = 2,grid_layout  = rbind(c(1, 1, 1, 1, 1, 2, 2),
                                                  c(1, 1, 1, 1, 1, 2, 2)),
  p1_theme = NULL,p2_theme = NULL,p1_guide = NULL,p2_guide = NULL
)
BetterVis_Barplot_Accumulative(
  data = iris, x_var = "Group", y1_var = "Petal.Width", y2_var = "percent", fill_var = "Group2",
  display = 3, barplot_color = barplot_color,
  ylim1 = c(0, 4.5), ylim2 = c(0, 1),
  axis_titles_x = "Cll Types", x_label_angle = 0, x_label_bold = TRUE, axis_title_x_size = 14,
  axis_titles_y1 = "Cell Count", axis_titles_y2 = "Sample Ratio",
  y1_label_bold = TRUE,y2_label_bold =  TRUE, axis_title_y1_size = 14,axis_title_y2_size = 14,
  title1 = "Cell Number Distribution", title2 = "Sample Ratio Distribution",
  title_size = 16,title_bold = TRUE,
  title1_size = 16, title1_bold = TRUE, title2_size = 16, title2_bold = TRUE,
  legend_show = TRUE, legend_title = "Legend", legend_size = 14, legend_position = "bottom",
  background_color = "white",
  coord_flip = TRUE,
  grid_ncol = 2,grid_layout  = rbind(c(1, 1, 1, 1, 1, 2, 2),
                                     c(1, 1, 1, 1, 1, 2, 2)),
  p1_theme = NULL,p2_theme = NULL,p1_guide = NULL,p2_guide = NULL
)

}
