% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Barplot_Bidirectional.R
\name{BetterVis_Barplot_Bidirectional}
\alias{BetterVis_Barplot_Bidirectional}
\title{Bidirectional Barplot Visualization Function}
\usage{
BetterVis_Barplot_Bidirectional(
  data,
  x_var,
  y_var,
  fill_var,
  barplot_color,
  x_label_angle = 0,
  x_label_bold = FALSE,
  axis_titles_x = "",
  axis_title_x_size = 12,
  y_label_bold = FALSE,
  axis_titles_y = "",
  y_label_angle = 0,
  axis_title_y_size = 12,
  title = "",
  title_size = 14,
  title_bold = FALSE,
  legend_show = TRUE,
  legend_title = "",
  legend_size = 10,
  background_color = "white",
  ylim = c(NA, NA),
  arrow = FALSE,
  plot_margin = unit(c(2, 0.3, 0.3, 0.3), "cm"),
  arrow_x1 = c(0.35, -0.2),
  arrow_x2 = c(0.48, 0.63),
  arrow_y1 = c(0.88, 0.88),
  arrow_y2 = c(0.88, 0.88),
  arrow_col = "black",
  arrow_size = 1,
  arrow_text = c("Stable Low DDS", "Stable High DDS"),
  arrow_text_size = 2,
  arrow_text_adjust = 0,
  right_text = FALSE,
  right_text1 = "",
  right_text2 = "",
  right_text_hjust = 1,
  right_text_vjust = 1,
  right_text_size = 12
)
}
\arguments{
\item{data}{A data frame containing the variables for plotting.}

\item{x_var}{A string specifying the variable for the x-axis (should be a factor variable).}

\item{y_var}{A string specifying the numeric variable for the y-axis.}

\item{fill_var}{A string specifying the variable used for fill colors (should be a factor variable).}

\item{barplot_color}{A vector specifying the fill colors corresponding to each level of the fill_var factor.}

\item{x_label_angle}{Numeric, indicating the angle for x-axis labels.}

\item{x_label_bold}{Logical, indicating if x-axis labels should be bold (default: \code{FALSE}).}

\item{axis_titles_x}{A string for the x-axis title (default: \code{""}).}

\item{axis_title_x_size}{Numeric, defining the font size of the x-axis title.}

\item{y_label_bold}{Logical, indicating if y-axis labels should be bold (default: \code{FALSE}).}

\item{axis_titles_y}{A string for the y-axis title (default: \code{""}).}

\item{y_label_angle}{Numeric, indicating the angle for y-axis labels.}

\item{axis_title_y_size}{Numeric, defining the font size of the y-axis title.}

\item{title}{Character, the main title of the plot (default: \code{""}).}

\item{title_size}{Numeric, defines the font size of the title (default: \code{14}).}

\item{title_bold}{Logical, indicating if the title should be bold (default: \code{FALSE}).}

\item{legend_show}{Logical, indicating if the legend should be shown (default: \code{TRUE}).}

\item{legend_title}{Character, the title for the legend (default: \code{""}).}

\item{legend_size}{Numeric, specifying the font size of the legend text.}

\item{background_color}{Character, the background color of the plot (default: \code{"white"}).}

\item{ylim}{A numeric vector to set the y-axis limits.}

\item{arrow}{Logical, whether to display arrows (default: \code{FALSE}).}

\item{plot_margin}{A \code{unit} object specifying the plot margin.}

\item{arrow_x1, }{arrow_x2, arrow_y1, arrow_y2 Numeric vectors defining the start and end points of arrows.}

\item{arrow_col}{Character, color of the arrows.}

\item{arrow_size}{Numeric, size of the arrows.}

\item{arrow_text}{A character vector of length 2, providing the text for labels adjacent to arrows.}

\item{arrow_text_size}{Numeric, size of the text adjacent to arrows.}

\item{arrow_text_adjust}{Numeric, vertical adjustment for text adjacent to arrows.}

\item{right_text}{Logical, whether to display text in the lower right (default: \code{FALSE}).}

\item{right_text1, }{right_text2 Characters, content for the text displayed in the lower right.}

\item{right_text_hjust, }{right_text_vjust Numeric adjustments for horizontal and vertical placement of right-side text.}

\item{right_text_size}{Numeric, size of the right-side text.}
}
\value{
A \code{ggplot} object with the customized bidirectional bar plot.
}
\description{
Generates bidirectional bar plots meant to display data comparison across categories with customized aesthetics and additional annotations to highlight specific data insights.
}
\details{
Creates bidirectional bar plots with extensive customization options. This function provides users with the ability to tailor themes, axes, titles, legends, and additional graphic elements such as arrows and annotations.
}
\examples{
library(BetterVis)
library(grid)

##  Data Input
data("BetterVis_Barplot_Bidirectional_example", package = "BetterVis")

## Simple Example
BetterVis_Barplot_Bidirectional(
data = BetterVis_Barplot_Bidirectional_example,x_var = "Percentage", y_var = "pathway", fill_var = "Species",
barplot_color = c("#ABDDDE", "#FAD510", "#C6CDF7", "#F4B5BD", "#0A9F9D", "#FAEED1", "#005295", "#E6A0C4", "#C52E19", "orange"),
title = "Species Contribution to Pathways",legend_title = "Species Contribution",
ylim = c(1, 8),
arrow = FALSE,plot_margin = unit(c(1,0.3,0.3,0.3),unit="cm"))



##  Advanced Example
BetterVis_Barplot_Bidirectional(
data = BetterVis_Barplot_Bidirectional_example,x_var = "Percentage", y_var = "pathway", fill_var = "Species",
barplot_color = c("#ABDDDE", "#FAD510", "#C6CDF7", "#F4B5BD", "#0A9F9D",
                    "#FAEED1", "#005295", "#E6A0C4", "#C52E19", "orange"),
axis_titles_x = "", x_label_angle = 0, x_label_bold = TRUE, axis_title_x_size = 14,
axis_titles_y = "", y_label_bold = TRUE, y_label_angle = 0,axis_title_y_size = 14,
title = "Species Contribution to Pathways", title_size = 16, title_bold = TRUE,
legend_show = TRUE, legend_title = "Species Contribution", legend_size = 10,
background_color = "white",
ylim = c(1, 8),
plot_margin = unit(c(2,0.3,0.3,0.3),unit="cm"),
arrow = TRUE,
arrow_x1 = c(0.35, 0.2), arrow_x2 = c(0.48, 0.63),
arrow_y1 = c(0.88, 0.88), arrow_y2 = c(0.88, 0.88),
arrow_col = "black", arrow_size = 0.5,
arrow_text = c("Enriched in Enriched in Disease Group","Enriched in Control Group"),arrow_text_size = 12,arrow_text_adjust = 0.03,
right_text = TRUE,right_text1 = "n = 154 for Disease",right_text2 = "n = 192 for Control",right_text_hjust = 1.2,right_text_vjust = -3,right_text_size =4
)

## We have modified the iris dataset to create these boxplots
iris$Group <- c(rep("Group2",75),rep("Group1",75))
iris$Group <-as.factor(iris$Group)
iris$Group2 <- rep(rep(paste0("Group", 1:10), each = 5),3)
iris$Group2 <-as.factor(iris$Group2)
BetterVis_Barplot_Bidirectional(
data = iris,x_var = "Petal.Width",  y_var = "Group2", fill_var = "Species",
barplot_color = c("#ABDDDE", "#FAD510", "#C6CDF7", "#F4B5BD", "#0A9F9D","#FAEED1", "#005295", "#E6A0C4", "#C52E19", "orange"),
axis_titles_x = "", x_label_angle = 0, x_label_bold = TRUE, axis_title_x_size = 14,
axis_titles_y = "", y_label_bold = TRUE, y_label_angle = 0,axis_title_y_size = 14,
title = "Species Contribution to Pathways", title_size = 16, title_bold = TRUE,
legend_show = TRUE, legend_title = "Species Contribution", legend_size = 10,
background_color = "white",
ylim = c(1, 10),
plot_margin = unit(c(2,2,1,0.5),unit="cm"),
arrow = TRUE,
arrow_x1 = c(0.35, 0.2), arrow_x2 = c(0.48, 0.63),
arrow_y1 = c(0.88, 0.88), arrow_y2 = c(0.88, 0.88),
arrow_col = "black", arrow_size = 0.5,
arrow_text = c("Enriched in stable low DDS","Enriched in stable high DDS"),arrow_text_size = 12,arrow_text_adjust = 0.03,right_text = TRUE,right_text1 = "n = 154 for stable high DDS",right_text2 = "n = 192 for stable low  DDS",right_text_hjust = 1.2,right_text_vjust = -6,right_text_size =4)
# +theme(axis.title.x = element_blank())+guides(fill=FALSE)
}
