% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Barplot_Bidirectional_example}
\alias{BetterVis_Barplot_Bidirectional_example}
\title{Bidirectional Barplot Example Dataset}
\format{
A data frame with 10 rows and 4 variables:
\describe{
  \item{pathway}{Character. Name of the biochemical pathway, indicating the specific pathway being referenced.}
  \item{group}{Character. Defines the enrichment group for the dataset, specifying categories such as \code{"Enriched in stable low DDS"}.}
  \item{Species}{Character. Represents the species involved, such as \code{"Bacteroides massiliensis"}, \code{"Escherichia coli"}, and others.}
  \item{Percentage}{Numeric. The percentage contribution of each species to the pathway, which can be both positive or negative to indicate directionality.}
}
}
\source{
Simulated data for visualization examples.
}
\usage{
BetterVis_Barplot_Bidirectional_example
}
\description{
A dataset designed to illustrate the use of bidirectional bar plots. It includes example data on different pathways, groupings, species, and their respective contribution percentages.
}
\keyword{datasets}
