% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Barplot_Circle.R
\name{BetterVis_Barplot_Circle}
\alias{BetterVis_Barplot_Circle}
\title{Circular Barplot Visualization Function}
\usage{
BetterVis_Barplot_Circle(
  data,
  x_var,
  y_var,
  fill_var,
  label_var = NULL,
  barplot_color,
  y_continuous_limit = c(-5, 6),
  point_shape = 19,
  point_alpha = 0.7,
  text_adjust = 0.4,
  text_size = 3,
  text_alpha = 1,
  text_bold = TRUE,
  legend_bold = TRUE,
  legend_size = 8,
  legend_title_size = 10,
  legend_title_bold = TRUE,
  size_legend = FALSE,
  fill_legend = TRUE
)
}
\arguments{
\item{data}{A data frame containing the variables for plotting.}

\item{x_var}{A string specifying the variable for the x-axis, which should be the grouping factor.}

\item{y_var}{A string specifying the numeric variable for the y-axis.}

\item{fill_var}{A string specifying the variable used for fill colors, which should be a factor variable.}

\item{label_var}{An optional string for the column that contains labels to annotate each bar.}

\item{barplot_color}{A vector specifying the colors corresponding to each level of the fill_var factor.}

\item{y_continuous_limit}{A numeric vector defining the limits for the y-axis scale, e.g., \code{c(-5, 6)}.}

\item{point_shape}{Numeric, defines the shape of the points on top of bars.}

\item{point_alpha}{Numeric, sets the transparency level for the points.}

\item{text_adjust}{Numeric, adjusts the distance of the labels from the bars.}

\item{text_size}{Numeric, defines the size of the bar labels.}

\item{text_alpha}{Numeric, sets the transparency level for the bar labels.}

\item{text_bold}{Logical, specifies whether bar labels should be bold (default: \code{TRUE}).}

\item{legend_bold}{Logical, specifies if the legend text should be bold (default: \code{TRUE}).}

\item{legend_size}{Numeric, specifies the size of the legend text.}

\item{legend_title_size}{Numeric, specifies the font size of the legend title.}

\item{legend_title_bold}{Logical, specifies if the legend title should be bold (default: \code{TRUE}).}

\item{size_legend}{Logical, determines if the legend for size should be shown (default: \code{FALSE}).}

\item{fill_legend}{Logical, determines if the legend for fill should be shown (default: \code{TRUE}).}
}
\value{
A \code{ggplot} object representing the circular bar plot visualization.
}
\description{
Generates a circular bar plot to visualize data distributions, offering customization of color, size, shape, and text for enhanced visual clarity and insight.
}
\details{
Creates circular bar plots with extensive customization options. The function allows users to adjust the appearance of bars, points, and labels, as well as control legend display and aesthetics.
}
\examples{
library(ggplot2)
library(dplyr)
library(ggtext)


##  示例数据1
data("BetterVis_Barplot_Circle_example", package = "BetterVis")
df <- BetterVis_Barplot_Circle_example
df <-df \%>\% select(geneSymbol,med_LFC,type,id)
df <- df \%>\% mutate(id =1:nrow(df))

## 示例数据2
barplot_color  <-  c("#FC5D5C", "#4B6A9D", "#167D7D")
BetterVis_Barplot_Circle(data = df,x_var = "id",y_var="med_LFC",fill_var = "type",label_var = "geneSymbol",
                         barplot_color = barplot_color)


BetterVis_Barplot_Circle(data = df, x_var = "id", y_var = "med_LFC", fill_var = "type",label_var = "geneSymbol",
                         barplot_color = barplot_color,
                         y_continuous_limit = c(-4, 3),
                         point_shape = 19,point_alpha = 0.7,
                         text_adjust = 0.5, text_size = 4,text_alpha = 0.9, text_bold = TRUE,
                         size_legend = FALSE, fill_legend = TRUE,
                         legend_bold = TRUE,legend_size = 10,legend_title_size = 12,legend_title_bold = TRUE)
#+theme(legend.position = "bottom") + guides(size = guide_legend(title = "Size Legend"))


}
