% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Barplot_Experiment.R
\name{BetterVis_Barplot_Experiment}
\alias{BetterVis_Barplot_Experiment}
\title{Experiment Barplot Visualization Function}
\usage{
BetterVis_Barplot_Experiment(
  data,
  x_var,
  y_var,
  fill_var,
  jitter = FALSE,
  jitter_width = 0.2,
  jitter_color = NULL,
  jitter_alpha = 0.7,
  jitter_size = 1.5,
  legend_show = TRUE,
  legend_title = NULL,
  barplot_color = NULL,
  facet = FALSE,
  facet_title_size = 16,
  comparison = NULL,
  comparison_col = NULL,
  ylim = NULL,
  title = NULL,
  x_label_angle = 45,
  x_label_bold = FALSE,
  axis_titles_x = "",
  axis_titles_y = "Y Axis",
  axis_title_bold = FALSE,
  y_label_bold = FALSE,
  background_color = "white",
  bar_border = TRUE,
  sig_type = "SYMBOL",
  error_bar_width = 0.2,
  error_bar_size = 1,
  bar_border_size = 1,
  custom_theme = NULL,
  custom_guides = NULL,
  comparison_method = "t.test"
)
}
\arguments{
\item{data}{A data frame containing the variables for plotting.}

\item{x_var}{A string specifying the variable for the x-axis, which should be a factor.}

\item{y_var}{A string specifying the numeric variable for the y-axis.}

\item{fill_var}{A string specifying the variable used for fill colors, which should also be a factor.}

\item{jitter}{Logical, indicating whether to display jittered points (default: \code{FALSE}).}

\item{jitter_width}{Numeric, defines the width of the jittered points (default: \code{0.2}).}

\item{jitter_color}{A vector specifying the colors for jittered points, matching the number of factor levels.}

\item{jitter_alpha}{Numeric, sets the transparency for jittered points (default: \code{0.7}).}

\item{jitter_size}{Numeric, specifies the size of the jittered points (default: \code{1.5}).}

\item{legend_show}{Logical, controls the display of the legend (default: \code{TRUE}).}

\item{legend_title}{A string for the legend title.}

\item{barplot_color}{A vector specifying fill colors for the bars, with a length equal to the number of factor levels.}

\item{facet}{Logical, indicating whether to use faceting in the plot (default: \code{FALSE}).}

\item{facet_title_size}{Numeric, size of the facet titles.}

\item{comparison}{A list of groups to be compared statistically.}

\item{comparison_col}{A string specifying the variable to be used for comparison.}

\item{ylim}{A numeric vector specifying the limits for the y-axis.}

\item{title}{A string for the plot title.}

\item{x_label_angle}{Numeric, angle for the x-axis labels.}

\item{x_label_bold}{Logical, specifies if the x-axis labels should be bold (default: \code{FALSE}).}

\item{axis_titles_x}{A string for the x-axis title, use "" if no title is desired.}

\item{axis_titles_y}{A string for the y-axis title, use "" if no title is desired.}

\item{axis_title_bold}{Logical, determines if axis titles should be bold (default: \code{FALSE}).}

\item{y_label_bold}{Logical, determines whether the y-axis labels should be bold (default: \code{FALSE}).}

\item{background_color}{A string for the background color, default is \code{"white"}.}

\item{bar_border}{Logical, indicates if a border should be added to the bars (default: \code{TRUE}).}

\item{sig_type}{A string indicating the type of significance mark to use (\code{"SYMBOL"} or \code{"NUMBER"}).}

\item{error_bar_width}{Numeric, width of the error bars.}

\item{error_bar_size}{Numeric, thickness of the error bars.}

\item{bar_border_size}{Numeric, thickness of the bar borders.}

\item{custom_theme, }{custom_guides Used for integration with additional theme or guide customization.}

\item{comparison_method}{A string specifying the statistical method for comparison (\code{"t.test"} or \code{"wilcox"}, default: \code{"t.test"}).}
}
\value{
A \code{ggplot} object representing the customized bar plot.
}
\description{
Provides functionality for generating bar plots with customizable aesthetics, error bars, jittering, and statistical comparisons, enabling detailed visual analysis of data.
}
\details{
This function creates bar plots with extensive options for customization, including features like jittered points, faceting, and statistical comparisons.
}
\examples{
library(ggpubr)
library(ggplot2)
library(RColorBrewer)
library(dplyr)
library(cowplot)

##  Data Input
data("BetterVis_Barplot_Experiment_example", package = "BetterVis")

##  Simple Example
BetterVis_Barplot_Experiment(
data = BetterVis_Barplot_Experiment_example,x_var = "supp",y_var = "len",fill_var = "dose",
comparison_col = "dose",comparison_method = "t.test",
comparison = list(c("dose_1", "dose_2"),c("dose_1", "dose_3"),c("dose_2", "dose_3")),
barplot_color = c("#E84A32", "#4CBCD2", "#00A18A"),
facet = TRUE,
ylim = c(0,45)
)

## Advanced Example

BetterVis_Barplot_Experiment(
data = BetterVis_Barplot_Experiment_example,x_var = "supp",y_var = "len",fill_var = "dose",
jitter = TRUE,jitter_color =c("black","black","black"),
jitter_width = 0.2,jitter_alpha = 0.7,jitter_size = 3,
legend_show = FALSE,legend_title = "Dose",
comparison_col = "dose",comparison_method = "t.test",
comparison = list(c("dose_1", "dose_2"),c("dose_1", "dose_3"),c("dose_2", "dose_3")),
barplot_color = c("#E84A32", "#4CBCD2", "#00A18A"),
facet = TRUE,facet_title_size = 16,
y_label_bold = TRUE,axis_titles_y = "",
x_label_angle = 45,x_label_bold = TRUE,axis_titles_x = c(""),
title = "",axis_title_bold = TRUE,
error_bar_width = 0.5,error_bar_size = 0.5,bar_border = TRUE,bar_border_size = 0.8,
background_color = "white",sig_type = "SYMBOL",ylim = c(0,45)
)+theme()+guides()


## We have modified the iris dataset to create these boxplots
iris$Group <- rep(rep(paste0("Group", 1:5), each = 10),3)
iris$Group <-as.factor(iris$Group)
BetterVis_Barplot_Experiment(
data = iris,x_var = "Group",y_var = "Petal.Width",fill_var = "Species",
jitter = TRUE,jitter_color =c("black","black","black"),
jitter_width = 0.2,jitter_alpha = 0.7,jitter_size = 3,
legend_show = FALSE,legend_title = "Dose",
comparison_col = "Species",comparison_method = "t.test",
comparison = list(c("setosa", "versicolor"),c("versicolor", "virginica"),c("setosa", "virginica")),
barplot_color = c("#E84A32", "#4CBCD2", "#00A18A"),
facet = TRUE,facet_title_size = 16,
y_label_bold = TRUE,axis_titles_y = "",
x_label_angle = 45,x_label_bold = TRUE,axis_titles_x = c("a"),
title = "",axis_title_bold = TRUE,
error_bar_width = 0.5,error_bar_size = 0.5,bar_border = TRUE,bar_border_size = 0.8,
background_color = "white",sig_type = "SYMBOL",ylim = c(0,4)
)

}
