% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Barplot_Horizontal.R
\name{BetterVis_Barplot_Horizontal}
\alias{BetterVis_Barplot_Horizontal}
\title{Horizontal Barplot Visualization Function}
\usage{
BetterVis_Barplot_Horizontal(
  data,
  x_var,
  y_var,
  fill_var,
  label_y_var,
  label_other_var,
  bar_width = 0.5,
  base_size = 35,
  label_y_size = 4.5,
  label_y_hjust = 0,
  label_y_vjust = 0,
  label_other_size = 4,
  label_other_hjust = 0,
  label_other_vjust = 2.5,
  label_y_start = 0.5,
  label_other_start = 0.5,
  fill_gradient = list(low = "#486b98", mid = "f5f2b1", high = "#b93735", midpoint = 3),
  color_gradient = list(low = "#486b98", mid = "#f5f2b1", high = "#b93735", midpoint = 3),
  fill_legend = TRUE,
  color_legend = FALSE,
  fill_legend_title,
  color_legend_title,
  legend_title_size = 15,
  legend_text_size = 15,
  xlab_text = "",
  ylab_text = "",
  plot_title = "",
  y_label_out = FALSE,
  label_other_display = TRUE
)
}
\arguments{
\item{data}{A data frame containing the variables for plotting.}

\item{x_var}{A string specifying the variable for the x-axis, which should be a factor.}

\item{y_var}{A string specifying the numeric variable for the y-axis.}

\item{fill_var}{A string specifying the variable used for fill colors.}

\item{label_y_var}{A string representing the column for y-axis labels.}

\item{label_other_var}{A string representing the column for additional labels within bars.}

\item{bar_width}{Numeric, specifies the width of the bars (default: \code{0.5}).}

\item{base_size}{Numeric, controls the scaling of the entire plot (default: \code{35}).}

\item{label_y_size}{Numeric, sets the size of the y-axis labels (default: \code{4.5}).}

\item{label_y_hjust}{Numeric, horizontal adjustment for y-axis labels (default: \code{0}).}

\item{label_y_vjust}{Numeric, vertical adjustment for y-axis labels (default: \code{0}).}

\item{label_other_size}{Numeric, size of additional labels (default: \code{4}).}

\item{label_other_hjust}{Numeric, horizontal adjustment for additional labels (default: \code{0}).}

\item{label_other_vjust}{Numeric, vertical adjustment for additional labels (default: \code{2.5}).}

\item{label_y_start}{Numeric, starting position for y-axis labels (default: \code{0.5}).}

\item{label_other_start}{Numeric, starting position for additional labels (default: \code{0.5}).}

\item{fill_gradient}{A list with "low", "mid", "high", "midpoint" for fill color gradient.}

\item{color_gradient}{A list with "low", "mid", "high", "midpoint" for text color gradient.}

\item{fill_legend}{Logical, determines if the legend for fill_var should be shown (default: \code{TRUE}).}

\item{color_legend}{Logical, determines if the legend for label_other_var should be shown (default: \code{FALSE}).}

\item{fill_legend_title}{The title for the fill legend.}

\item{color_legend_title}{The title for the color legend.}

\item{legend_title_size}{Numeric, size of the legend titles (default: \code{15}).}

\item{legend_text_size}{Numeric, size of the legend text (default: \code{15}).}

\item{xlab_text}{The label for the x-axis.}

\item{ylab_text}{The label for the y-axis.}

\item{plot_title}{The title for the plot.}

\item{y_label_out}{Logical, if \code{TRUE}, y-axis labels are placed outside (default: \code{FALSE}).}

\item{label_other_display}{Logical, if \code{TRUE}, additional labels are displayed inside bars (default: \code{TRUE}).}
}
\value{
A \code{ggplot} object presenting the horizontal barplot visualization.
}
\description{
Creates horizontal bar plots that allow for rich customization of visual elements, such as label positioning, color gradients, and legend design, enabling detailed and informative visual representations of data.
}
\details{
This function generates horizontal bar plots with extensive customization options for controlling bar widths, label positions, color gradients, and legends.
}
\examples{
library(ggplot2)
library(dplyr)

##  示例数据1
data("BetterVis_Barplot_Horizontal_example", package = "BetterVis")
## Simple Example

BetterVis_Barplot_Horizontal(
data = BetterVis_Barplot_Horizontal_example,x_var = "Description",y_var = "Count",fill_var = "LogP",
label_y_var = "Description", label_other_var = "Genes",fill_legend_title = "logP",
fill_gradient = list(low = "#486b98", mid = "#f5f2b1", high = "#b93735", midpoint = 3),
y_label_out = TRUE,label_other_display = TRUE
)


## Advanced Example
BetterVis_Barplot_Horizontal(
data = BetterVis_Barplot_Horizontal_example,x_var = "Description",y_var = "Count",fill_var = "LogP",
  label_y_var = "Description",label_other_var = "Genes",
  bar_width = 0.4,base_size = 30,
  label_y_start = 0.5,label_y_size = 5,label_y_hjust = 0,label_y_vjust = 1,
  label_other_start = 0.5,label_other_size = 4.5,label_other_hjust = 0,label_other_vjust = 2.4,
  fill_gradient = list(low = "#486b98", mid = "#f5f2b1", high = "#b93735", midpoint = 3),
  color_gradient = list(low = "#486b98", mid = "#f5f2b1", high = "#b93735", midpoint = 3),
  fill_legend = TRUE, fill_legend_title="logp",color_legend = FALSE,color_legend_title = "col_lengend",
  legend_title_size = 16, legend_text_size = 16,
  xlab_text = " Names",ylab_text = "Gene Count", plot_title = "KEGG Enrichment Analysis",
  y_label_out = TRUE,label_other_display = TRUE
)
##+theme(axis.title.x = element_blank())
}
