% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Boxplot_Experiment.R
\name{BetterVis_Boxplot_Experiment}
\alias{BetterVis_Boxplot_Experiment}
\title{BetterVis Boxplot Function}
\arguments{
\item{data}{A data frame.}

\item{x_var}{A string specifying the x-axis variable.}

\item{y_var}{A string specifying the y-axis variable.}

\item{fill_var}{A string specifying the fill variable.}

\item{use_facet}{Logical. Whether to use facets in the plot.}

\item{title}{Optional. The title of the plot.}

\item{x_label_angle}{Numeric. Angle of x-axis labels (default: 45 degrees).}

\item{x_label_bold}{Logical. Whether to bold the x-axis labels.}

\item{y_label_bold}{Logical. Whether to bold the y-axis labels.}

\item{axis_titles}{Character vector. Titles for x and y axes.}

\item{axis_title_bold}{Logical. Whether to bold axis titles.}

\item{custom_colors}{Optional. Custom colors for the fill variable.}

\item{jitter}{Logical. Whether to add jitter to the points.}

\item{jitter_color}{Optional. Custom colors for jitter points.}

\item{jitter_width}{Numeric. Width of jitter (default: 0.1).}

\item{jitter_alpha}{Numeric. Transparency of jitter points (default: 1).}

\item{jitter_size}{Numeric. Size of jitter points (default: 1).}

\item{legend_show}{Logical. Whether to show the legend (default: TRUE).}

\item{legend_title}{Optional. Title of the legend.}

\item{box_width}{Numeric. Width of the boxplots (default: 0.75).}

\item{comparison_col}{Character. Column used for statistical comparison.}

\item{comparison}{List. Comparisons for statistical testing.}

\item{comparison_method}{Character. Statistical test method ("wilcox" or "t.test").}

\item{box_background}{Logical. Whether to include a background for the boxplot.}

\item{jitter_legend}{Logical. Whether jitter points should appear in the legend.}

\item{fill_legend}{Logical. Whether the fill variable should appear in the legend.}

\item{background_color}{Character. Color of the plot background.}

\item{background_border}{Logical. Whether to include a border around the background.}

\item{sig_type}{Character. Type of significance label ("SYMBOL" or "p.format").}

\item{guides_modifications}{Optional. Custom modifications to guides.}

\item{violin}{Logical. Whether to include a violin layer (default: FALSE).}

\item{...}{Additional arguments passed to ggplot2 layers.}
}
\value{
A ggplot object.
}
\description{
Creates enhanced boxplots with customization options for ggplot2.
}
\examples{

library(BetterVis)

##示例数据1
data("BetterVis_Boxplot_Experiment_example", package = "BetterVis")
iris$Group <- rep(rep(paste0("Group", 1:5), each = 10),3)
iris$Group <-as.factor(iris$Group)

## Simple Example
BetterVis_Boxplot_Experiment(
  data=BetterVis_Boxplot_Experiment_example,
  x_var = "supp", y_var = "len", fill_var = "dose",
  use_facet = TRUE,
  custom_colors = c("#FC5D5C", "#4B6A9D", "#167D7D"),
  comparison_col = "dose",
  comparison = list(c("dose_1","dose_2"),c("dose_1","dose_3"),c("dose_2","dose_3")),
  ylim = c(0,40)
)

## Advanced Example
BetterVis_Boxplot_Experiment(
  data=BetterVis_Boxplot_Experiment_example, x_var = "supp", y_var = "len", fill_var = "dose",
  use_facet = TRUE, title = "",
  x_label_angle = 45, x_label_bold = TRUE,
  y_label_bold = TRUE, axis_titles = c("X Axis", "Y Axis"),
  axis_title_bold = TRUE,
  custom_colors = c("#FC5D5C", "#4B6A9D", "#167D7D"),
  jitter = TRUE,
  jitter_color = c("#FC5D5C", "#4B6A9D", "#167D7D"),
  jitter_width = 1,jitter_alpha = 1,jitter_size = 2,
  legend_show = FALSE, legend_title = "",
  box_width = 0.6,
  comparison_col = "dose",
  comparison = list(c("dose_1","dose_2"),c("dose_1","dose_3"),c("dose_2","dose_3")),
  comparison_method = "wilcox",
  box_background = TRUE,background_color = "white",background_border = TRUE,sig_type = "SYMBOL",violin = FALSE,ylim = c(0,40)
)+theme()+guides()

## We have modified the iris dataset to create these boxplots
iris$Group <- rep(rep(paste0("Group", 1:5), each = 10),3)
iris$Group <-as.factor(iris$Group)

BetterVis_Boxplot_Experiment(
  iris, x_var = "Group", y_var = "Petal.Width", fill_var = "Species",
  use_facet = TRUE,
  title = "",
  x_label_angle = 45,
  x_label_bold = TRUE,
  y_label_bold = TRUE,
  axis_titles = c("X Axis", "Y Axis"),
  axis_title_bold = TRUE,
  custom_colors = c("#FC5D5C", "#4B6A9D", "#167D7D"),
  jitter = TRUE,
  jitter_color = c("#FC5D5C", "#4B6A9D", "#167D7D"),
  jitter_width = 1,
  jitter_alpha = 1,
  jitter_size = 2,
  legend_show = TRUE,
  legend_title = "",
  box_width = 0.6,
  comparison_col = "Species",
  comparison = list(c("setosa", "versicolor"),c("setosa", "virginica"),c("versicolor", "virginica")),
  comparison_method = "wilcox",
  box_background = TRUE,background_color = "white",background_border = TRUE,sig_type = "NUMBER",violin = FALSE,ylim = c(0,3.5)
)
#+theme(axis.title.x = element_blank())+guides(fill=FALSE) #'



}
