% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Circos_Ringplot.R
\name{BetterVis_Circos_Ringplot}
\alias{BetterVis_Circos_Ringplot}
\title{Create a Circular Dendrogram (Ring Plot)}
\usage{
BetterVis_Circos_Ringplot(data, root, color = NULL)
}
\arguments{
\item{data}{A data frame containing the hierarchical data. It must include columns named "trait", "SNP", and "Gene".}

\item{root}{A character string specifying the name of the root node for the dendrogram.}

\item{color}{(Optional) A vector of colors for the main branches (`trait`). If `NULL`, colors will be automatically generated.}
}
\value{
A `ggraph` object representing the circular dendrogram.
}
\description{
This function visualizes hierarchical data as a circular dendrogram. It takes a
data frame with columns representing different levels of a hierarchy (e.g., trait,
SNP, Gene), constructs a graph, and plots it using `ggraph`.
}
\examples{
if (requireNamespace("ggraph", quietly = TRUE) && requireNamespace("tidygraph", quietly = TRUE)) {
  # Load example data from the package
  data("BetterVis_Circos_Ringplot_example", package = "BetterVis")

  # Generate the plot
  BetterVis_Circos_Ringplot(data = BetterVis_Circos_Ringplot_example, root = "Cough", color = NULL)
}
}
