% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Color.R
\name{BetterVis_Color}
\alias{BetterVis_Color}
\title{BetterVis Color Function}
\usage{
BetterVis_Color(
  type = c("discrete", "continuous"),
  odd_or_even = NULL,
  paired = FALSE,
  pair_num = NULL,
  n = NULL,
  option = 1,
  style = c("rcolorbrewer", "Science", "Cell", "Lancet")
)
}
\arguments{
\item{type}{A character string specifying whether to use "discrete" or "continuous" color schemes. Default is \code{NULL}.}

\item{odd_or_even}{A character string specifying whether the total number of colors should be "odd" or "even" (only applicable to discrete types). Default is \code{NULL}.}

\item{paired}{A logical value indicating whether to use paired colors. Default is \code{FALSE}. If \code{TRUE}, colors will be repeated in groups based on \code{pair_num}.}

\item{pair_num}{An integer specifying the number of colors in each group when \code{paired = TRUE}. This must be specified when using paired colors.}

\item{n}{An integer specifying the total number of colors needed. Required for both discrete and continuous types.}

\item{option}{An integer specifying the color scheme option. Options are \code{1}, \code{2}, or \code{3}.}

\item{style}{A character string specifying the color style to use. Options are \code{"Science"}, \code{"Nature"}, \code{"Cell"}, or \code{"Lancet"}.}

\item{...}{Additional arguments passed to \code{ggplot2} layers (if applicable).}
}
\value{
A vector of color values, either discrete or continuous, depending on the specified parameters.
}
\description{
Provides discrete and continuous color schemes for visualizations with advanced customization options.
}
\details{
This function provides enhanced color customization options for creating plots. It supports both discrete and continuous color palettes,
with several predefined styles such as "Nature", "Science", "Cell", and "Lancet". The function allows for specifying odd or even numbers
of colors, pairing options, and custom schemes.
}
\examples{
# 离散型非配对取色
BetterVis_Color(type = "discrete", odd_or_even = "even",  n = 10, style = "Science", option = 1)
# 离散型配对取色
BetterVis_Color(type = "discrete", odd_or_even = "even",  n = 9, style = "Science", option = 2,paired = TRUE, pair_num = 3)
# 连续型取色
BetterVis_Color(type = "continuous", n = 9, style = "Science", option = 1)
}
