% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Dotplot_Beeswarm_example}
\alias{BetterVis_Dotplot_Beeswarm_example}
\title{Example Data for a Beeswarm Plot}
\format{
A data frame with columns including:
\describe{
  \item{Type}{A categorical variable for the x-axis, representing vehicle type.}
  \item{MPG.city}{A numeric variable for the y-axis, representing city MPG.}
}
}
\source{
Simulated data for visualization examples, originally from a CSV file.
}
\usage{
BetterVis_Dotplot_Beeswarm_example
}
\description{
A dataset containing city miles per gallon (MPG) for various vehicle types,
suitable for creating a beeswarm plot with `BetterVis_Dotplot_Beeswarm`.
}
\keyword{datasets}
