% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Dotplot_Mixbar_example}
\alias{BetterVis_Dotplot_Mixbar_example}
\title{Example Data for a Mixed Dot Plot and Bar Plot}
\format{
A data frame with columns including:
\describe{
  \item{species}{A character string indicating the penguin species.}
  \item{bill_length_mm}{A numeric variable for the x-axis.}
  \item{bill_depth_mm}{A numeric variable for the y-axis.}
  \item{body_mass_g}{A numeric variable for point size.}
}
}
\source{
Data originally from the `palmerpenguins` package, by way of a CSV file.
}
\usage{
BetterVis_Dotplot_Mixbar_example
}
\description{
A dataset containing measurements for different species of penguins, suitable
for creating a scatter plot with `BetterVis_Dotplot_Mixbar`. This data is
derived from the `palmerpenguins` package.
}
\keyword{datasets}
