% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Dotplot_Multifacet_example}
\alias{BetterVis_Dotplot_Multifacet_example}
\title{Example Data for a Multi-Faceted Dot Plot}
\format{
A data frame with 150 rows and 5 columns:
\describe{
  \item{Sepal.Length}{Numeric. Length of the sepal.}
  \item{Sepal.Width}{Numeric. Width of the sepal.}
  \item{Petal.Length}{Numeric. Length of the petal.}
  \item{Petal.Width}{Numeric. Width of the petal.}
  \item{Species}{Factor. The species of iris.}
}
}
\source{
The base R `iris` dataset, by way of a CSV file.
}
\usage{
BetterVis_Dotplot_Multifacet_example
}
\description{
A dataset containing the classic `iris` measurements, suitable for creating
a matrix of scatter plots with `BetterVis_Dotplot_Multifacet`.
}
\keyword{datasets}
