% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Heatmap_Bicontinuous_example}
\alias{BetterVis_Heatmap_Bicontinuous_example}
\title{Example Data for a Bi-Continuous Heatmap}
\format{
A data frame with columns:
\describe{
  \item{ID}{A variable for the x-axis.}
  \item{name}{A variable for the y-axis.}
  \item{value}{A continuous variable for the fill color.}
}
}
\source{
Simulated data for visualization examples, originally from a CSV file.
}
\usage{
BetterVis_Heatmap_Bicontinuous_example
}
\description{
A dataset in long format suitable for creating a tile-based heatmap with
`BetterVis_Heatmap_Bicontinuous`.
}
\keyword{datasets}
