% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Heatmap_Circle.R
\name{BetterVis_Heatmap_Circle}
\alias{BetterVis_Heatmap_Circle}
\title{Create a Circular Heatmap}
\usage{
BetterVis_Heatmap_Circle(
  data,
  value_var,
  classified_text_var,
  classified_value_var,
  heatmap_color = alpha(met.brewer("Hiroshige"), 1),
  text_size = 4,
  barwidth = 9,
  barheight = 0.5,
  legend_break = 1,
  legend_text_size = 12,
  classified_value_var_label_size = 10
)
}
\arguments{
\item{data}{A data frame containing the plotting data in a long format.}

\item{value_var}{A character string for the column name of the numeric variable for the tile fill color.}

\item{classified_text_var}{A character string for the column name of the categorical variable for the outer ring labels.}

\item{classified_value_var}{A character string for the column name of the categorical variable for the inner rings (y-axis).}

\item{heatmap_color}{A vector of colors for the color gradient. Defaults to a `MetBrewer` palette.}

\item{text_size}{Numeric. The size for the curved text labels. Default is `4`.}

\item{barwidth, barheight}{Numeric. The width and height of the color bar legend.}

\item{legend_break}{Numeric. The interval for the legend's color scale breaks. Default is `1`.}

\item{legend_text_size}{Numeric. The font size for the legend text. Default is `12`.}

\item{classified_value_var_label_size}{Numeric. The font size for the radial axis labels. Default is `10`.}
}
\value{
A `ggplot` object representing the circular heatmap.
}
\description{
This function generates a circular heatmap (radial tile plot) using `ggplot2`
and `geomtextpath`. It is designed to visualize tabular data in a polar
coordinate system with curved text labels.
}
\examples{
 library(ggplot2)
 library(MetBrewer)
 library(geomtextpath)
 library(dplyr)
## Data Input
data("BetterVis_Heatmap_Circle_example",package="BetterVis")
BetterVis_Heatmap_Circle(data = BetterVis_Heatmap_Circle_example, value_var = "value", classified_text_var = "id", classified_value_var = "name")
BetterVis_Heatmap_Circle(data = BetterVis_Heatmap_Circle_example, value_var = "value", classified_text_var = "id", classified_value_var = "name",
heatmap_color = alpha(met.brewer("Hiroshige"), 1),
text_size =4,barwidth = 9, barheight = 0.5,legend_break = 1,legend_text_size=12,
classified_value_var_label_size = 10)

## We have modified the iris dataset to create these boxplots
iris$Group <- rep(paste0("Group", 1:10), times = 15)
iris$Group <-as.factor(iris$Group)
iris$Group2 <- rep(rep(paste0("Color", 1:5), each = 10),3)
iris$Group2 <- sample(iris$Group2)
iris$Group2 <-as.factor(iris$Group2)
BetterVis_Heatmap_Circle(data = iris, value_var = "Sepal.Width", classified_text_var = "Group", classified_value_var = "Group2")
}
