% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Heatmap_Circle_example}
\alias{BetterVis_Heatmap_Circle_example}
\title{Example Data for a Circular Heatmap}
\format{
A data frame with columns:
\describe{
  \item{value}{A continuous variable for the fill color.}
  \item{id}{A categorical variable for the outer ring labels.}
  \item{name}{A categorical variable for the inner rings.}
}
}
\source{
Simulated data for visualization examples, originally from a CSV file.
}
\usage{
BetterVis_Heatmap_Circle_example
}
\description{
A dataset in long format suitable for creating a circular heatmap with
`BetterVis_Heatmap_Circle`.
}
\keyword{datasets}
