% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Heatmap_Common.R
\name{BetterVis_Heatmap_Common}
\alias{BetterVis_Heatmap_Common}
\title{Create a Classic Heatmap}
\usage{
BetterVis_Heatmap_Common(
  data,
  heatmap_color = MetBrewer::met.brewer("Hiroshige"),
  legend_position = "top",
  legend_title = NULL,
  legend_break = 1,
  legend_text_size = 10,
  legend_text_bold = TRUE,
  x_text_size = 12,
  x_text_bold = TRUE,
  x_text_angle = 90,
  y_text_size = 12,
  y_text_bold = TRUE,
  trans = FALSE
)
}
\arguments{
\item{data}{A data frame or matrix for plotting. Rows and columns will become the heatmap axes.}

\item{heatmap_color}{A vector of colors for the color gradient. Defaults to a `MetBrewer` palette.}

\item{legend_position}{Character string. The position of the legend (e.g., "top", "right"). Default is `"top"`.}

\item{legend_title}{Character string for the legend title.}

\item{legend_break}{Numeric. The interval for the legend's color scale breaks. Default is `1`.}

\item{legend_text_size}{Numeric. The font size for the legend text.}

\item{legend_text_bold}{Logical. If `TRUE`, legend text is bold.}

\item{x_text_size, y_text_size}{Numeric. Font size for x and y axis text.}

\item{x_text_bold, y_text_bold}{Logical. If `TRUE`, axis text is bold.}

\item{x_text_angle}{Numeric. The angle of x-axis text.}

\item{trans}{Logical. If `TRUE`, the input data is transposed before plotting. Default is `FALSE`.}
}
\value{
A `ggplot` object representing the heatmap.
}
\description{
This function generates a classic heatmap from a data frame or matrix using
`ggplot2`. It reshapes wide data into a long format and provides extensive
options for aesthetic customization.
}
\examples{
  # Load example data
  library(magrittr)
  library(MetBrewer)
  library(ggplot2)
  data("BetterVis_Heatmap_Common_example", package = "BetterVis")

  # Example 1: Basic usage
  BetterVis_Heatmap_Common(BetterVis_Heatmap_Common_example, trans = FALSE)

  # Example 2: More customized plot
BetterVis_Heatmap_Common(BetterVis_Heatmap_Common_example, trans = FALSE,
heatmap_color = met.brewer("Hiroshige"),legend_break = 1,
legend_position = "top", legend_title = "",legend_text_size = 8,legend_text_bold = TRUE,
x_text_size = 12,x_text_bold = TRUE,x_text_angle = 90,
y_text_size = 12,y_text_bold = TRUE)
#+guides(fill="none")
#+theme(axis.x.text=element_text(size=15)
}
