% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Heatmap_Common_example}
\alias{BetterVis_Heatmap_Common_example}
\title{Example Data for a Common Heatmap}
\format{
A data frame with numeric values.
}
\source{
Simulated data for visualization examples, originally from a CSV file.
}
\usage{
BetterVis_Heatmap_Common_example
}
\description{
A numeric data frame in wide format suitable for creating a heatmap with
`BetterVis_Heatmap_Common`. Rows and columns represent different features
or samples, and the cells contain numeric values.
}
\keyword{datasets}
