% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Heatmap_Mantel.R
\name{BetterVis_Heatmap_Mantel}
\alias{BetterVis_Heatmap_Mantel}
\title{Create a Mantel Test Heatmap}
\usage{
BetterVis_Heatmap_Mantel(
  phenotype_data,
  expression_data,
  spec_select,
  significant_symbol = TRUE,
  heatmap_color = RColorBrewer::brewer.pal(11, "PuOr"),
  line_size = c(0.5, 1, 2),
  line_color = (scales::hue_pal())(3),
  phenotype_text_size = 5
)
}
\arguments{
\item{phenotype_data}{A data frame containing phenotype data (or the first matrix).}

\item{expression_data}{A data frame or matrix containing expression data (or the second matrix for the heatmap).}

\item{spec_select}{A character vector of column names from `phenotype_data` to be used in the Mantel test against `expression_data`.}

\item{significant_symbol}{Logical. If `TRUE`, adds significance asterisks to the correlation heatmap. Default is `TRUE`.}

\item{heatmap_color}{A vector of colors for the heatmap gradient. Defaults to a palette from `RColorBrewer`.}

\item{line_size}{A numeric vector of length 3 for the line widths corresponding to Mantel's r categories.}

\item{line_color}{A vector of 3 colors for the lines corresponding to Mantel's p-value categories.}

\item{phenotype_text_size}{Numeric. The font size for the labels of the phenotype variables.}
}
\value{
A `ggplot` object.
}
\description{
This function creates a visualization that combines a correlation heatmap of one
dataset with the results of a Mantel test linking it to a second dataset. It
heavily relies on the `linkET` package.
}
\examples{
library(linkET)
library(dplyr)
library(ggplot2)
library(tidyverse)
library(RColorBrewer)
library(vegan)
if (requireNamespace("linkET", quietly = TRUE)) {
  # Load example data
  data("BetterVis_Heatmap_Mantel_example1", package = "BetterVis")
  data("BetterVis_Heatmap_Mantel_example2", package = "BetterVis")

  # Example 1: Basic usage
  BetterVis_Heatmap_Mantel(
    phenotype_data = BetterVis_Heatmap_Mantel_example2,
    expression_data = BetterVis_Heatmap_Mantel_example1,
    spec_select = c("stage", "age", "gender", "CA125", "CEA", "CA15.3")
  )

  # Example 2: More customized plot
  BetterVis_Heatmap_Mantel(
    phenotype_data = BetterVis_Heatmap_Mantel_example2,
    expression_data = BetterVis_Heatmap_Mantel_example1,
    spec_select = c("stage", "age", "gender", "CA125", "CEA", "CA15.3"),
    significant_symbol = TRUE,
    heatmap_color = brewer.pal(11, "RdYlBu"),
    line_size = c(0.5, 1, 1.5),
    line_color = c("blue", "green", "grey"),
    phenotype_text_size = 5
  ) +
  theme(legend.text = element_text(size = 16), legend.title = element_text(size = 18))
}
}
