% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Heatmap_Mantel_example1}
\alias{BetterVis_Heatmap_Mantel_example1}
\title{Example Expression Data for Mantel Test Heatmap}
\format{
A data frame with numeric values.
}
\source{
Simulated data for visualization examples.
}
\usage{
BetterVis_Heatmap_Mantel_example1
}
\description{
An example numeric data frame representing expression data, suitable for the
`expression_data` argument in `BetterVis_Heatmap_Mantel`.
}
\keyword{datasets}
