% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Heatmap_Mantel_example2}
\alias{BetterVis_Heatmap_Mantel_example2}
\title{Example Phenotype Data for Mantel Test Heatmap}
\format{
A data frame with various clinical indicators.
}
\source{
Simulated data for visualization examples.
}
\usage{
BetterVis_Heatmap_Mantel_example2
}
\description{
An example data frame representing clinical or phenotype data, suitable for the
`phenotype_data` argument in `BetterVis_Heatmap_Mantel`.
}
\keyword{datasets}
