% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Heatmap_Multi_example}
\alias{BetterVis_Heatmap_Multi_example}
\title{Example Data for a Multi-Layered Heatmap}
\format{
A data frame with columns for sample IDs, numeric values, and
  categorical annotations.
\describe{
  \item{case.ID}{Character. A unique identifier for each sample on the x-axis.}
  \item{Primary Metastasis}{Character. An example annotation variable.}
  \item{Consensus classification}{Character. An example annotation variable.}
  \item{UNC classification}{Character. An example annotation variable.}
  \item{...}{Other columns are numeric variables for the main heatmap.}
}
}
\source{
Simulated data for visualization examples, originally from a CSV file.
}
\usage{
BetterVis_Heatmap_Multi_example
}
\description{
A dataset containing a mix of numeric and categorical variables suitable for
creating a composite heatmap with `BetterVis_Heatmap_Multi`.
}
\keyword{datasets}
