% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Network_PPI_edges}
\alias{BetterVis_Network_PPI_edges}
\title{Example Edge Data for a PPI Network Plot}
\format{
A data frame with 'from', 'to', and 'weight' columns.
}
\source{
Simulated data for visualization examples.
}
\usage{
BetterVis_Network_PPI_edges
}
\description{
An example data frame representing edge information for a network plot,
suitable for the `edges_data` argument in `BetterVis_Network_PPI`.
}
\keyword{datasets}
