% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Pie_Multi_example_out}
\alias{BetterVis_Pie_Multi_example_out}
\title{Example Outer Pie Data for Multi Pie Plot}
\format{
A data frame with columns for positioning, grouping, and proportions.
}
\source{
Simulated data for visualization examples.
}
\usage{
BetterVis_Pie_Multi_example_out
}
\description{
An example dataset in long format suitable for the `data_out` argument of
`BetterVis_Pie_Multi`, defining the outer pie chart slices.
}
\keyword{datasets}
