% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Pie_Single.R
\name{BetterVis_Pie_Single}
\alias{BetterVis_Pie_Single}
\title{Create a Single Pie Chart}
\usage{
BetterVis_Pie_Single(
  df,
  ID_var,
  Proportion_var,
  colors = NULL,
  label = FALSE,
  size_adjust = 1
)
}
\arguments{
\item{df}{A data frame containing the data for plotting.}

\item{ID_var}{A character string for the column name of the categorical variable that defines the pie slices.}

\item{Proportion_var}{A character string for the column name of the numeric variable that defines the slice proportions.}

\item{colors}{(Optional) A vector of colors for the pie slices. If `NULL`, a default color palette is generated.}

\item{label}{Logical. If `TRUE`, adds labels with names and percentages directly onto the pie slices. Default is `FALSE`.}

\item{size_adjust}{A single numeric value to uniformly scale text elements (labels and legend). Default is `1`.}
}
\value{
A `ggplot` object representing the pie chart.
}
\description{
This function generates a customizable pie chart from a data frame using
`ggplot2`. It allows for custom colors and optional in-slice labels.
}
\examples{
# Load example data
data("BetterVis_Pie_Single_example", package = "BetterVis")

# Define a color palette
col <- c("#A2D4AB", "#82B3C9", "#E0BBE4", "#FAD0C3", "#C9E4DE", "#FFECB3")

BetterVis_Pie_Single(df = BetterVis_Pie_Single_example,ID_var = "Cancer",Proportion_var = "Percent",
                    colors = col,label = TRUE ,size_adjust = 1)+theme()+guides()


}
