% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Ridgeline_Facet.R
\name{BetterVis_Ridgeline_Facet}
\alias{BetterVis_Ridgeline_Facet}
\title{Create a Faceted Ridgeline Plot with Optional Annotations}
\usage{
BetterVis_Ridgeline_Facet(
  data,
  x_var,
  y_var,
  facet_var = NULL,
  nrow = NULL,
  ncol = NULL,
  annotation_var = NULL,
  color,
  annotation = TRUE,
  color_annotation_var = c("#68d0ce", "#f4a4c9", "#4955d0", "#e09a2a", "#de6a73"),
  all_text_adjust = 1,
  combined_width = c(1.5, 0.8)
)
}
\arguments{
\item{data}{A data frame containing the plotting data.}

\item{x_var}{A character string specifying the column name for the numeric variable (x-axis).}

\item{y_var}{A character string specifying the column name for the categorical variable (y-axis).}

\item{facet_var}{(Optional) A character string for the column name used to create plot facets.}

\item{nrow}{(Optional) The number of rows for the facet layout.}

\item{ncol}{(Optional) The number of columns for the facet layout.}

\item{annotation_var}{(Optional) A character string for the column name in `data` used for the annotation plot's fill colors.}

\item{color}{A vector of colors for the ridgeline fills, with a length equal to the number of levels in `y_var`.}

\item{annotation}{A logical value. If `TRUE`, an annotation plot is added. Default is `TRUE`.}

\item{color_annotation_var}{A vector of colors for the annotation plot.}

\item{all_text_adjust}{A single numeric value to uniformly scale text elements. Default is `1`.}

\item{combined_width}{A numeric vector of length 2 specifying the relative widths of the ridgeline plot and the annotation plot. Default is `c(1.5, 0.8)`.}
}
\value{
A `ggplot` or `cowplot` object.
}
\description{
This function generates a faceted ridgeline plot using the `ggridges` package.
It allows for visualizing distributions across different categories, further
subdivided into facets, and supports an optional annotation sidebar.
}
\examples{
# Load example data from the package
data("BetterVis_Ridgeline_Facet_example", package = "BetterVis")

# Define a color palette for the ridgelines
sampleColor <- c('#F6EB5E','#CD8280','#F0F0F0', '#54D0B4',
                 '#D2D1D1','#78C3ED', '#E69F00', '#CBDFF3',
                 '#3BFFB8', '#ECAD67','#B3B2B2',
                 '#959595', '#E98FBD','#2672B2')

# Example 1: Faceting by a single variable
BetterVis_Ridgeline_Facet(
  data = BetterVis_Ridgeline_Facet_example,
  x_var = "CytoTRACEScore",
  y_var = "patient",
  facet_var = "smoking1",
  annotation_var = "mutationGroup",
  color = sampleColor
)

# Example 2: Faceting with specified rows and columns
BetterVis_Ridgeline_Facet(
  data = BetterVis_Ridgeline_Facet_example,
  x_var = "CytoTRACEScore",
  y_var = "patient",
  facet_var = "Group",
  nrow = 2,
  ncol = 3,
  color = sampleColor,
  annotation = TRUE,
  annotation_var = "mutationGroup",
  color_annotation_var = c("#68d0ce", "#f4a4c9", "#4955d0", "#e09a2a", "#de6a73"),
  all_text_adjust = 1.1,
  combined_width = c(1.5, 0.8)
)
}
