% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Ridgeline_Facet_example}
\alias{BetterVis_Ridgeline_Facet_example}
\title{Example Data for a Faceted Ridgeline Plot}
\format{
A data frame with columns including:
\describe{
  \item{CytoTRACEScore}{A numeric variable for the x-axis.}
  \item{patient}{A categorical variable for the y-axis, defining the ridges.}
  \item{smoking1}{A categorical variable for faceting.}
  \item{Group}{Another categorical variable for faceting.}
  \item{mutationGroup}{An additional categorical variable for annotation.}
}
}
\source{
Simulated data for visualization examples, originally from a CSV file.
}
\usage{
BetterVis_Ridgeline_Facet_example
}
\description{
A dataset containing numeric scores and multiple categorical groupings
suitable for creating a faceted ridgeline plot with
`BetterVis_Ridgeline_Facet`.
}
\keyword{datasets}
