% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Ridgeline_Single_example}
\alias{BetterVis_Ridgeline_Single_example}
\title{Example Data for a Single Ridgeline Plot}
\format{
A data frame with columns including:
\describe{
  \item{CytoTRACEScore}{A numeric variable for the x-axis.}
  \item{patient}{A categorical variable for the y-axis, defining the ridges.}
  \item{mutationGroup}{An additional categorical variable for annotation.}
}
}
\source{
Simulated data for visualization examples, originally from a CSV file.
}
\usage{
BetterVis_Ridgeline_Single_example
}
\description{
A dataset containing numeric scores and categorical groupings suitable for
creating a ridgeline plot with `BetterVis_Ridgeline_Single`.
}
\keyword{datasets}
