% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Sankey_Single_example}
\alias{BetterVis_Sankey_Single_example}
\title{Example Data for a Single Flow Sankey Diagram}
\format{
A data frame with multiple rows and 4 columns:
\describe{
  \item{AgeGroup}{Factor. The age category of the individual.}
  \item{Diagnosis}{Factor. The diagnosis received by the individual.}
  \item{Treatment}{Factor. The treatment administered.}
  \item{Outcome}{Factor. The final outcome for the individual.}
}
}
\source{
Simulated data for visualization examples.
}
\usage{
BetterVis_Sankey_Single_example
}
\description{
A dataset where each row represents an individual observation's flow through
a series of categorical states. This data is suitable for creating a Sankey
diagram with `BetterVis_Sankey_Single`.
}
\keyword{datasets}
