% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{BetterVis_Venn_UpsetR_example}
\alias{BetterVis_Venn_UpsetR_example}
\title{Example Data for UpSet Plot}
\format{
A named list with multiple elements, each being a character vector.
}
\source{
Simulated data for visualization examples, originally stored in an .rds file.
}
\usage{
BetterVis_Venn_UpsetR_example
}
\description{
A named list of character vectors, where each vector represents a set of
elements. This dataset is used to demonstrate the functionality of
`BetterVis_Venn_UpsetR`.
}
\keyword{datasets}
