% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{Boxplot_Experiment_example}
\alias{Boxplot_Experiment_example}
\title{Boxplot Example Data}
\format{
A data frame with 60 rows and 3 variables:
\describe{
  \item{len}{Numeric. Length of the sample.}
  \item{supp}{Character. Type of supplement (\code{"VC"} or \code{"OJ"}).}
  \item{dose}{Factor. Dose level of the supplement (\code{"dose_1"}, \code{"dose_2"}, or \code{"dose_3"}).}
}
}
\source{
Simulated example data.
}
\usage{
Boxplot_Experiment_example
}
\description{
A dataset to demonstrate boxplot functionality. It includes measurements
for different supplement types and doses.
}
\keyword{datasets}
