% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Boxplot_MixViolin.R
\name{BetterVis_Boxplot_MixViolin}
\alias{BetterVis_Boxplot_MixViolin}
\title{BetterVis Boxplot MixViolin Function}
\usage{
BetterVis_Boxplot_MixViolin(
  data,
  x_var,
  y_var,
  fill_color,
  box_size = 0.5,
  title = "",
  title_size = 16,
  title_bold = FALSE,
  title_hjust = 0.5,
  title_vjust = 1,
  axis_x_title = element_text(size = 22, color = "#b35959", face = "bold"),
  axis_y_title = element_text(size = 22, color = "#b35959", face = "bold"),
  xlab = NULL,
  ylab = NULL,
  border = TRUE,
  panel_background_color = "#f3f6f6",
  legend_show = TRUE,
  legend_title = NULL,
  legend_title_size = 15,
  legend_title_bold = FALSE,
  legend_text_size = 12,
  legend_text_bold = FALSE,
  stat_compare_means = TRUE,
  compare_symbol = "SYMBOL",
  y_max_adjust = 1,
  y_min_adjust = 1
)
}
\arguments{
\item{data}{A data frame containing the variables for plotting.}

\item{x_var}{A string specifying the variable for the x-axis (should be a factor variable).}

\item{y_var}{A string specifying the numeric variable for the y-axis.}

\item{fill_color}{A vector of colors corresponding to each level of the x-axis variable.}

\item{title}{Character. The main title of the plot (default: \code{""}).}

\item{title_size}{Numeric. Defines the font size of the title (default: \code{16}).}

\item{title_bold}{Logical. Defines whether the title should be bold (default: \code{FALSE}).}

\item{title_hjust}{Numeric. Adjusts the horizontal alignment of the title (default: \code{0.5}).}

\item{title_vjust}{Numeric. Adjusts the vertical alignment of the title (default: \code{1}).}

\item{axis_x_title}{A \code{ggplot2::element_text} object specifying x-axis title styling.}

\item{axis_y_title}{A \code{ggplot2::element_text} object specifying y-axis title styling.}

\item{xlab}{Character. The label for the x-axis.}

\item{ylab}{Character. The label for the y-axis.}

\item{border}{Logical. Whether to display a border around the plot (default: \code{TRUE}).}

\item{panel_background_color}{Character. Color of the panel background (default: \code{"#f3f6f6"}).}

\item{legend_show}{Logical. Whether to show the legend (default: \code{TRUE}).}

\item{legend_title}{Character. Title for the legend (default: \code{NULL}).}

\item{legend_title_size}{Numeric. Size of the legend title text (default: \code{15}).}

\item{legend_title_bold}{Logical. Whether the legend title should be bold (default: \code{FALSE}).}

\item{legend_text_size}{Numeric. Size of the legend text (default: \code{12}).}

\item{legend_text_bold}{Logical. Whether the legend text should be bold (default: \code{FALSE}).}

\item{stat_compare_means}{Logical. Whether to add statistical comparisons (default: \code{TRUE}).}

\item{compare_symbol}{Character. Whether to display comparison results as symbols ("SYMBOL") or numeric p-values ("NUMBER") (default: \code{"SYMBOL"}).}

\item{y_max_adjust}{Numeric. Adjusts the upper limit of the y-axis (default: \code{1}).}

\item{y_min_adjust}{Numeric. Adjusts the lower limit of the y-axis (default: \code{1}).}

\item{...}{Additional arguments passed to \code{ggplot2} layers.}
}
\value{
A \code{ggplot} object representing the customized violin-boxplot visualization.
}
\description{
Generates a violin-boxplot hybrid visualization with extensive customization options, statistical comparisons, and color controls.
}
\details{
Creates an enhanced mixed visualization combining violin plots, boxplots, and scatter points with extensive customization options.
The function allows users to apply different themes, customize axes, add statistical comparisons, and define aesthetic properties.
}
\examples{
# Load necessary libraries
library(BetterVis)
library(dplyr)
data("BetterVis_Boxplot_MixViolin_example", package = "BetterVis")
barplot_color = c("#ABDDDE", "#FAD510", "#C6CDF7", "#F4B5BD", "#FAEED1","#0A9F9D",  "#005295", "#E6A0C4","#C52E19","orange")

## Simple Example
# Example 1: Simple visualization
BetterVis_Barplot_Accumulative(
data = BetterVis_Barplot_Accumulative_example, x_var = "celltype", y1_var = "cell_num", y2_var = "percent", fill_var = "Patient",
display = 3, barplot_color = barplot_color,
ylim1 = c(0, 15000), ylim2 = c(0, 1),
coord_flip = TRUE
)


## Advanced Example
# Example 2: Customized visualization with titles and axes
BetterVis_Barplot_Accumulative(
data = BetterVis_Barplot_Accumulative_example, x_var = "celltype", y1_var = "cell_num", y2_var = "percent", fill_var = "Patient",
display = 3, barplot_color = barplot_color,
ylim1 = c(0, 15000), ylim2 = c(0, 1),
axis_titles_x = "Cell Types", x_label_angle = 0, x_label_bold = TRUE, axis_title_x_size = 14,
axis_titles_y1 = "Cell Count", axis_titles_y2 = "Sample Ratio",
y1_label_bold = TRUE,y2_label_bold =  TRUE, axis_title_y1_size = 14,axis_title_y2_size = 14,
title1 = "Cell Number Distribution", title2 = "Sample Ratio Distribution",
title_size = 10,title_bold = TRUE,
legend_show = TRUE, legend_title = "Legend", legend_size = 14, legend_position = "bottom", legend_row = NULL,legend_col=NULL,
background_color = "white",
coord_flip = TRUE,
grid_ncol = 2,grid_layout  = rbind(c(1, 1, 1, 1, 1, 2, 2),
                                     c(1, 1, 1, 1, 1, 2, 2)),
p1_theme = NULL,p2_theme = NULL,p1_guide = NULL,p2_guide = NULL
)

##  We have modified the iris dataset to create these boxplots
iris$Group <- rep(paste0("Group", 1:10), times = 15)
iris$Group <-as.factor(iris$Group)
iris$Group2 <- rep(rep(paste0("Color", 1:5), each = 10),3)
iris$Group2 <- sample(iris$Group2)
iris$Group2 <-as.factor(iris$Group2)
iris<-iris\%>\%dplyr::select(Petal.Width,Group2,Group)
iris<-distinct(iris,Group,Group2,.keep_all = TRUE)
iris = plyr::ddply(iris,'Group',transform,percent=Petal.Width/sum(Petal.Width))
BetterVis_Barplot_Accumulative(
data = iris, x_var = "Group", y1_var = "Petal.Width", y2_var = "percent", fill_var = "Group2",
display = 3, barplot_color = barplot_color,
ylim1 = c(0, 4.5), ylim2 = c(0, 1),
axis_titles_x = "Cell Types", x_label_angle = 0, x_label_bold = TRUE, axis_title_x_size = 14,
axis_titles_y1 = "Cell Count", axis_titles_y2 = "Sample Ratio",
y1_label_bold = TRUE,y2_label_bold =  TRUE, axis_title_y1_size = 14,axis_title_y2_size = 14,
title1 = "Cell Number Distribution", title2 = "Distribution",
title_size = 16,title_bold = TRUE,
title1_size = 16, title1_bold = TRUE, title2_size = 16, title2_bold = TRUE,
legend_show = TRUE, legend_title = "Legend", legend_size = 14, legend_position = "bottom",
background_color = "white",
coord_flip = TRUE,
grid_ncol = 2,grid_layout  = rbind(c(1, 1, 1, 1, 1, 2, 2),
                                     c(1, 1, 1, 1, 1, 2, 2)),
p1_theme = NULL,p2_theme = NULL,p1_guide = NULL,p2_guide = NULL
)



}
