% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterVis_Heatmap_Bicontinuous.R
\name{BetterVis_Heatmap_Bicontinuous}
\alias{BetterVis_Heatmap_Bicontinuous}
\title{Create a Heatmap for Bi-Continuous Data}
\usage{
BetterVis_Heatmap_Bicontinuous(
  data,
  x_var,
  y_var,
  fill_var,
  vline_position = 0.25,
  hline_position = 0,
  vline_color = "black",
  hline_color = "black",
  hline_linetype = "dashed",
  heatmap_color = rev(MetBrewer::met.brewer("Hiroshige")),
  legend_break = 1,
  legend_title = NULL,
  legend_position = "bottom",
  legend_title_position = "bottom"
)
}
\arguments{
\item{data}{A data frame containing the data for plotting.}

\item{x_var, y_var}{Character strings for the column names of the x and y-axis variables.}

\item{fill_var}{Character string for the column name of the continuous variable for the tile fill color.}

\item{vline_position, hline_position}{Numeric. The position for vertical and horizontal reference lines.}

\item{vline_color, hline_color}{Character strings. The color of the vertical and horizontal lines.}

\item{hline_linetype}{Character string. The line type for the horizontal line (e.g., "dashed", "solid").}

\item{heatmap_color}{A vector of colors to be used for the color gradient. Defaults to a `MetBrewer` palette.}

\item{legend_break}{Numeric. The interval for the legend's color scale breaks. Default is `1`.}

\item{legend_title}{Character string for the legend title. Supports markdown via `ggtext`.}

\item{legend_position}{Character string. The position of the legend (e.g., "bottom", "right").}

\item{legend_title_position}{Character string. The position of the legend title relative to the legend.}
}
\value{
A `ggplot` object representing the heatmap.
}
\description{
This function generates a heatmap (tile plot) from data in a long format,
where two variables map to the x and y axes, and a third continuous variable
maps to the fill color.
}
\examples{
if (requireNamespace("MetBrewer", quietly = TRUE) && requireNamespace("ggtext", quietly = TRUE)) {
  # Load example data
  library(ggplot2)
  data("BetterVis_Heatmap_Bicontinuous_example", package = "BetterVis")
  library(MetBrewer)
  # Example 1: Basic usage
  BetterVis_Heatmap_Bicontinuous(
    data = BetterVis_Heatmap_Bicontinuous_example,
    x_var = "ID", y_var = "name", fill_var = "value",
    legend_title = "Change"
  )

  # Example 2: More customized plot
  BetterVis_Heatmap_Bicontinuous(
    data = BetterVis_Heatmap_Bicontinuous_example,
    x_var = "ID", y_var = "name", fill_var = "value",
    vline_position = 0.5, hline_position = 5,
    vline_color = "blue", hline_color = "red", hline_linetype = "solid",
    heatmap_color = rev(met.brewer("Isfahan1")),
    legend_break = 2,
    legend_title = "**Fold Change**", # Example of markdown title
    legend_position = "right",
    legend_title_position = "top"
  ) +
  theme(axis.text = element_text(size = 12, face = "bold"))
}
}
